Create march_denrell.py
This script creates a simple learning model similar to the one in Denrell and March (2001)
The code presents just the main result of that paper. In specific, that agents end up believing that an option with variance has a lower value than an option without variance. Because of this, the stable option ends up being chosen more.  This phenomenon can be seen in figure 1. I additionally plot the average payoff on every period of the simulation, this shows a funnel shape as in later periods agents choose the stable option more and more. I print the expected value of the option and the probability of an agent choosing the stable option at the end of the simulation.

Additionally, I run a second simulation to show how large the differences in means has to be to overcome the Hotstove effect. 
In the specific case shown, one needs to change the mean of the variable arm from 0.5 to over 0.62 to have around 50% of the last choice being the variable arm. 

Many changes can and have been made to this model. You can come up with your own.

References:
Denrell, J., & March, J. G. (2001). Adaptation as information restriction: The hot stove effect. Organization Science, 12(5), 523-538.